// src/app/models/vendedores.ts
export class Vendedores {
  VEN_CODIGO: string;
  VEN_NOMBRE: string;
  VEN_COMISION1: number;
  VEN_COMISION2: number;
  VEN_CLAVE: string;
  CAJ_CODIGO: string;
  BOD_CODIGO: string;
  COM_CODIGO: string;
  VEN_DIRECCION1: string;
  VEN_DIRECCION2: string;
  VEN_TELEFONO1: string;
  VEN_TELEFONO2: string;
  VEN_FAX: string;
  VEN_CORREO: string;
  VEN_ZONA: string;
  VEN_PROVINCIA: string;
  CIU_CODIGO: string;
  VEN_JEFEZONA: string;
  VEN_IDE: string;
  CON_CODIGO: string;
  VEN_FECHAINGRESO: Date;
  VEN_FECHANACIMIENTO: Date;
  VEN_TRANSPORTE: string;
  VEN_OBSERVACIONES: string;

  constructor() {}
}